requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.0
pillow>=10.0.0
opencv-python>=4.8.0
pytest>=7.4.0
pydantic-settings>=2.0.0
tenacity>=8.2.0
cachetools>=5.3.0
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tldextract>=5.1.0
pyzbar>=0.1.9
streamlit>=1.30.0
requests>=2.31.0
